:sparkles: rename package names and add doc.go
