+++ internal/k8s/benchmark/job/job.go
@@ -14,7 +14,7 @@
 // limitations under the License.
 //
 
-package benchmark
+package job
 
 import (
 	"context"
