+++ internal/k8s/benchmark/operator/operator.go
@@ -14,7 +14,7 @@
 // limitations under the License.
 //
 
-package benchmark
+package operator
 
 import (
 	"context"
